services/ai_openai_service/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
openai>=1.68,<2.0
python-dotenv>=1.0,<2.0
